[server/routeRaspberry] 라즈베리파이 네트워크 통신 라우팅 분리
