+++ server/server.js
@@ -21,6 +21,7 @@ app.use(bodyParser.urlencoded({ extended: true }));     // body-parser의 기본
 // 라우팅
 app.use('/', require('./routes/home'));
 app.use('/arduino', require('./routes/arduino'));
+app.use('/rasp', require('./routes/rasp'));
 app.use('/mobile', require('./routes/mobile/mobile'));
 app.use('/mobile/vending', require('./routes/mobile/vending'));
 app.use('/mobile/drink', require('./routes/mobile/drink'));
